+++ MasonAzureTest/Models/Special.cs
@@ -5,6 +5,6 @@ namespace MasonAzureTest.Models
     public class Special
     {
         public List<TrollyQuantity> Quantities { get; set; }
-        public long Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
